+++ CHANGELOG.md
@@ -33,6 +33,9 @@
 
 ### Removed
 
+- [#5077](https://github.com/ChainSafe/forest/pull/5077) Remove
+  `peer_tipset_epoch` from the metrics.
+
 ### Fixed
 
 ## Forest v0.23.1 "Lappe"